requirements.txt
appdirs==1.4.3
javaobj-py3==0.2.3
packaging==16.8
pyparsing==2.2.0
six==1.10.0
